requirements.txt
streamlit
pytesseract
PyMuPDF
groq
python-dotenv
pandas
pdf2image
Pillow